% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonfig.R
\name{fillList}
\alias{fillList}
\title{Recursively apply glue on each list element}
\usage{
fillList(l, .open, .close)
}
\arguments{
\item{l}{:list}

\item{.open}{:character\cr Opening delimiter}

\item{.close}{:character\cr Closing delimiter}
}
\value{
:list\cr The full list with glued element
}
\description{
Recursively apply glue on each list element
}
