% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonfig.R
\name{get}
\alias{get}
\title{Get given config}
\usage{
get(value = NULL, .open = "{", .close = "}", ...)
}
\arguments{
\item{value}{:character\cr Name of value (NULL to read all values)}

\item{.open}{:character\cr Opening delimiter}

\item{.close}{:character\cr Closing delimiter}

\item{...}{Same arguments as \link[config:get]{config::get}}
}
\value{
:list \cr A list with \emph{config awared} glued element
}
\description{
Read from the currently active configuration as would do config::get and
offers you the power of glue with no extra fees.
}
\details{
For additional details see the package website at
 \href{https://github.com/supermayo/gonfig}{https://github.com/supermayo/gonfig}.
}
\seealso{
\code{\link[config:get]{config::get}},
\code{\link[glue:glue_data]{glue::glue_data}}
}
