% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{rglue}
\alias{rglue}
\title{Recursive glue}
\usage{
rglue(string, env, .open, .close)
}
\arguments{
\item{string}{:Maybe character\cr Maybe a character string.}

\item{env}{:list\cr An environment, list or data frame used to lookup values.}

\item{.open}{:character\cr Opening delimiter}

\item{.close}{:character\cr Closing delimiter}
}
\value{
:character\cr A glued string.
}
\description{
Recursively glue the string using environment until convergence.
}
